Generic PHY workaround problem
The laird scripts check if the PHY is generic, and if it is, then they
skip setting up the interface. In our case, the PHY drive is listed as
generic so that means eth0 is not setup. This commit removes the generic
phy checks so that the setup is always done.