Workaround for laird sdk settings wipe-out on reboot
The laird sdk has a bug that is triggered when the system starts without
a profiles.conf file. So we start with a default profiles.conf in order to
work around the bug. The default profiles.conf was taken by starting a
system without a profiles.conf file and rebooting without making any other
changes.